+++ src/app/views/base/printcustomers/printcustomers.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import {ClientService} from '../../../services/client.service';
 import {Users} from '../../../models/Users';
 import { NgToastService } from 'ng-angular-popup';
+import { NgxBootstrapConfirmService } from 'ngx-bootstrap-confirm';
 
 @Component({
   selector: 'app-navs',
@@ -11,7 +12,8 @@ import { NgToastService } from 'ng-angular-popup';
 export class PrintcustomersComponent {
   customers: any[]=[];
 
-  constructor(private clientService: ClientService, private toast: NgToastService) { 
+  constructor(private clientService: ClientService, private toast: NgToastService,
+    private ngxComfirmService: NgxBootstrapConfirmService) { 
 
   this.getAllClient();
   }
@@ -31,34 +33,40 @@ export class PrintcustomersComponent {
   }
 
   deleteCustomer(id:number){
-    this.clientService.delete(id).subscribe({
-      next: data=>{
-        if(data.status == 200){
 
-          this.toast.success({
-            detail:"Customer deleted",
-            summary:data.body.message,
-            duration: 3000
-           });
-           this.getAllClient();
-
-        }
-        
-      },
-      error: error=>{
-        if(error.status == 404){
-
-          this.toast.warning({
-            detail:"Le client n'existe pas!!!",
-            summary:error.body.message,
-            duration: 3000
-           });
-        }
-
-        
+    this.ngxComfirmService.confirm({
+      title:'Voulez vous effacer cette client?',
+      confirmLabel: 'Okay',
+      declineLabel: 'Cancel'
+    }).then((res: boolean) => {
+      if (res) {
+        this.clientService.delete(id).subscribe({
+          next: data=>{
+            if(data.status == 200){
+    
+              this.toast.success({
+                detail:"Customer deleted",
+                summary:data.body.message,
+                duration: 3000
+               });
+               this.getAllClient();
+            }
+            
+          },
+          error: error=>{
+            if(error.status == 404){
+    
+              this.toast.warning({
+                detail:"Le client n'existe pas!!!",
+                summary:error.body.message,
+                duration: 3000
+               });
+            } 
+          }
+        });
+      } else {
+        console.log('Cancel');
       }
-
-      
     });
   }
